Change types from `Int32` to `Int64`
Totals could be quite large, at least in theory.

+++ src/haystack/transaction/totals.cr
@@ -3,10 +3,10 @@ struct Haystack::Transaction::Totals
 
   getter pending_transfers : Int64
   getter pending_transfers_by_currency : Array(Amount)
-  getter total_transactions : Int32
+  getter total_transactions : Int64
   getter total_volume : Int64
   getter total_volume_by_currency : Array(Amount)
-  getter unique_customers : Int32?
+  getter unique_customers : Int64?
 
   struct Item
     include Response
